requirements.txt
# ai-service/requirements.txt (MINIMAL - Essential packages only)

# FastAPI and server
fastapi
uvicorn[standard]
python-multipart

# LLM APIs
google-generativeai
anthropic

# Document processing
pypdf
python-docx
openpyxl
python-pptx
markdown

# ChromaDB for vector storage
chromadb

# Data processing
pandas
numpy

# Cloud storage
boto3

# Authentication
PyJWT

# Data validation
pydantic
pydantic-settings

# HTTP client
httpx

# Text processing and embeddings
sentence-transformers

# Environment
python-dotenv

# Additional document processing
beautifulsoup4
lxml

# Development tools (optional)
pytest
black
isort